chore: remove ColorHelper from Uno.UWP and fix build
