+++ src/Uno.UWP/Uno.Tests.csproj
@@ -27,12 +27,4 @@
 		<ProjectReference Include="..\Uno.UI.Dispatching\Uno.UI.Dispatching.Tests.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<Compile Include="..\Uno.UI\UI\Input\PointerPoint.cs" />
-		<Compile Include="..\Uno.UI\UI\Input\PointerPointProperties.cs" />
-		<Compile Include="..\Uno.UI\UI\Input\PointerPointPropertiesExtensions.cs" />
-		<Compile Include="..\Uno.UI\UI\Input\PointerUpdateKind.cs" />
-		<Compile Include="..\Uno.UI\UI\Colors.cs" />
-		<Compile Include="..\Uno.UI\UI\Text\FontWeights.cs" />
-	</ItemGroup>
 </Project>
