+++ src/Uno.UWP/Uno.Wasm.csproj
@@ -26,12 +26,4 @@
 		<ProjectReference Include="..\Uno.UI.Dispatching\Uno.UI.Dispatching.Wasm.csproj" />
 	</ItemGroup>
 
-	<ItemGroup>
-		<Compile Include="..\Uno.UI\UI\Input\PointerPoint.cs" />
-		<Compile Include="..\Uno.UI\UI\Input\PointerPointProperties.cs" />
-		<Compile Include="..\Uno.UI\UI\Input\PointerPointPropertiesExtensions.cs" />
-		<Compile Include="..\Uno.UI\UI\Input\PointerUpdateKind.cs" />
-		<Compile Include="..\Uno.UI\UI\Colors.cs" />
-		<Compile Include="..\Uno.UI\UI\Text\FontWeights.cs" />
-	</ItemGroup>
 </Project>
